Rename things, fix styles
