+++ lib/containers/issueish-timeline-container.js
@@ -7,9 +7,9 @@ import Commits from './timeline-items/commits-container.js';
 import IssueComment from './timeline-items/issue-comment-container.js';
 import MergedEvent from './timeline-items/merged-event-container.js';
 import HeadRefForcePushedEvent from './timeline-items/head-ref-force-pushed-event-container.js';
-import CommitCommentThread from './timeline-items/commit-comment-thread';
+import CommitCommentThread from './timeline-items/commit-comment-thread-container';
 
-function collectionRenderer(Component, styleAsContainer = true) {
+function collectionRenderer(Component, styleAsTimelineItem = true) {
   return class GroupedComponent extends React.Component {
     static displayName = `Grouped(${Component.name})`
 
@@ -25,7 +25,7 @@ function collectionRenderer(Component, styleAsContainer = true) {
     }
 
     render() {
-      return <div className={styleAsContainer ? 'timeline-item' : ''}>{this.props.nodes.map(this.renderNode)}</div>;
+      return <div className={styleAsTimelineItem ? 'timeline-item' : ''}>{this.props.nodes.map(this.renderNode)}</div>;
     }
 
     @autobind
@@ -45,7 +45,7 @@ function collectionRenderer(Component, styleAsContainer = true) {
 const timelineItems = {
   Commit: Commits,
   CommitCommentThread: collectionRenderer(CommitCommentThread, false),
-  IssueComment: collectionRenderer(IssueComment),
+  IssueComment: collectionRenderer(IssueComment, false),
   MergedEvent: collectionRenderer(MergedEvent),
   HeadRefForcePushedEvent: collectionRenderer(HeadRefForcePushedEvent),
 };
