+++ lib/containers/timeline-items/issue-comment-container.js
@@ -15,7 +15,7 @@ export class IssueComment extends React.Component {
   render() {
     const comment = this.props.item;
     return (
-      <div className="issue">
+      <div className="issue timeline-item">
         <div className="info-row">
           <Octicon className="pre-timeline-item-icon" icon="comment" />
           <img className="author-avatar" src={comment.author.avatarUrl} title={comment.author.login} />
